encapsulate styles
