+++ src/components/Map.tsx
@@ -7,8 +7,9 @@
  * ------------------------------------------
  */
 
-import * as React   from 'react';
-import { css }      from '@emotion/react';
+import * as React from 'react';
+
+import * as Styles from 'styles/components/Map.styles';
 
 import MapController from 'MapAPI';
 import { loader }    from 'index';
@@ -55,17 +56,8 @@ class Map extends React.Component<MapProps> {
 
   render()  {
 
-    const mapStyle = css`
-    
-      width: 100%;
-      height: 720px;
-      border: 1px solid black;
-      border-radius: 5px;
-
-    `;
-
     return (
-      <div id="map" css={mapStyle}></div>
+      <div id="map" css={Styles.map}></div>
     );
 
   }
