+++ src/components/MiniModal.tsx
@@ -7,10 +7,10 @@
  * ------------------------------------------
  */
 
-import React   from 'react';
-import { css } from '@emotion/react';
+import React from 'react';
 
-import { modalBaseStyles } from 'GlobalStyles';
+import { modalBase } from 'styles/global.styles';
+import * as Styles   from 'styles/components/MiniModal.styles';
 
 /* ------------------------------------------
  *
@@ -38,123 +38,28 @@ class MiniModal extends React.Component<ModalProps> {
   
   render() {
 
-    /* -----------------------
-     *    COMPONENT STYLES
-     * -----------------------
-     */
-
-    const containerStyle = css`
-    
-      display: flex;
-      flex-direction: column;
-      top: 25%;
-      z-index: 3;
-      width: 250px;
-      height: fit-content;
-      min-width: 250px;
-      min-height: auto;
-    
-    `;
-
-    const textStyle = css`
-    
-      text-align: center;
-      border: 1px solid black;
-      padding: 25px;
-      margin: 5px;
-      border-radius: 8px;
-    
-    `;
-
-    const footerStyle = css`
-    
-      display: flex;
-      justify-content: space-between;
-      margin-left: 20px;
-      margin-right: 20px;
-      padding-bottom: 20px;
-    
-    `;
-
-    const genericButtonStyle = css`
-    
-      height: 50px;
-      border-radius: 5px;
-    
-    `;
-
-    const yesButtonStyle = css`
-
-      ${genericButtonStyle};
-      width: fit-content;
-      min-width: 55px;
-      margin: 3px;
-      border: 2px solid rgb(208, 208, 208);
-      background-color: white;
-
-      &:hover {
-        background-color: lightgreen;
-      }
-    
-    `;
-
-    const noButtonStyle = css`
-    
-      ${genericButtonStyle};
-      width: fit-content;
-      min-width: 55px;
-      margin: 3px;
-      border: 2px solid rgb(208, 208, 208);
-      background-color: white;
-
-      &:hover {
-        background-color: salmon;
-      }
-    
-    `;
-
-    const tertiaryButtonStyle = css`
-    
-      ${genericButtonStyle};
-      width: fit-content;
-      min-width: 55px;
-      margin: 3px;
-      border: 2px solid rgb(208, 208, 208);
-      background-color: white;
-
-      &:hover {
-        background-color: lightgrey;
-      }
-    
-    `;
-
-    /* -----------------------
-     *       HTML CONTENT
-     * -----------------------
-     */
-
     let message       = this.props.content.message;
     let confirmText   = this.props.content.confirmText;
     let closeText     = this.props.content.closeText;
     let tertiaryText  = this.props.content.tertiaryText;
 
-    let confirmButton   = (confirmText  ? <button onClick={this.props.actionPrimary} type="button" className="yes-button" css={yesButtonStyle}>{confirmText}</button>             : null);
-    let closeButton     = (closeText    ? <button onClick={this.props.actionSecondary} type="button" className="no-button" css={noButtonStyle}>{closeText}</button>                   : null);
-    let tertiaryButton  = (tertiaryText ? <button onClick={this.props.actionTertiary} type="button" className="tertiary-button" css={tertiaryButtonStyle}>{tertiaryText}</button> : null);
+    let confirmButton   = (confirmText  ? <button onClick={this.props.actionPrimary} type="button" className="yes-button" css={Styles.yesButton}>{confirmText}</button>             : null);
+    let closeButton     = (closeText    ? <button onClick={this.props.actionSecondary} type="button" className="no-button" css={Styles.noButton}>{closeText}</button>               : null);
+    let tertiaryButton  = (tertiaryText ? <button onClick={this.props.actionTertiary} type="button" className="tertiary-button" css={Styles.tertiaryButton}>{tertiaryText}</button> : null);
 
     return (
-      <div className="modal-backdrop confirm-modal" css={modalBaseStyles.modalBackdrop}>
-        <div className="modal-container confirm-modal" css={[modalBaseStyles.modalContainer, containerStyle]}>
+      <div className="modal-backdrop confirm-modal" css={modalBase.modalBackdrop}>
+        <div className="modal-container confirm-modal" css={[modalBase.modalContainer, Styles.container]}>
 
           <div className="modal-body confirm-modal">
-            <div className="modal-text confirm-modal" css={textStyle}>
+            <div className="modal-text confirm-modal" css={Styles.text}>
 
               {message}
 
             </div> 
           </div>
 
-          <div className="modal-footer confirm-modal" css={footerStyle}>
+          <div className="modal-footer confirm-modal" css={Styles.footer}>
 
             {confirmButton}
             {closeButton}
