+++ src/components/misc/ConfirmText.tsx
@@ -1,5 +1,3 @@
-/** @jsxImportSource @emotion/react */
-
 /* ------------------------------------------
  *
  *                  IMPORT
@@ -11,31 +9,10 @@ import React from 'react';
 
 /* ------------------------------------------
  *
- *                   MISC
+ *                CONFIRMTEXT
  * 
  * ------------------------------------------
  */
-
-type TrashProps = app.component.misc.trashButton.Props;
-
-class TrashButton extends React.Component<TrashProps> {
-
-  /*  Description:
-   *    Renders a trash button. Calls a handleTrash function on click.
-   *
-   *  Expects props:
-   *    - handleTrash => Function. Handle the functionality for trash button click.
-   */
-
-  render() {
-
-    return (
-      <button onClick={this.props.handleTrash} type="button" name="trash-button" className="trash-button"><i className="far fa-trash-alt"></i></button>
-    );
-
-  }
-
-}
 
 type ConfirmProps = app.component.misc.confirmText.Props;
 
@@ -59,4 +36,4 @@ class ConfirmText extends React.Component<ConfirmProps> {
 
 }
 
-export { TrashButton, ConfirmText };
+export default ConfirmText;
