+++ src/styles/components/Map.styles.ts
@@ -1,0 +1,35 @@
+/** @jsxImportSource @emotion/react */
+
+/* ------------------------------------------
+ *
+ *                  IMPORT
+ * 
+ * ------------------------------------------
+ */
+
+import { css } from '@emotion/react';
+
+/* ------------------------------------------
+ *
+ *         COMPONENT "MAP" STYLES
+ * 
+ * ------------------------------------------
+ */
+
+const map = css`
+    
+  width: 100%;
+  height: 720px;
+  border: 1px solid black;
+  border-radius: 5px;
+
+`;
+
+/* ------------------------------------------
+ *
+ *                  EXPORT
+ * 
+ * ------------------------------------------
+ */
+
+export { map };
