+++ src/styles/components/ModalButtons.styles.ts
@@ -1,0 +1,34 @@
+/** @jsxImportSource @emotion/react */
+
+/* ------------------------------------------
+ *
+ *                  IMPORT
+ * 
+ * ------------------------------------------
+ */
+
+import { css } from '@emotion/react';
+
+/* ------------------------------------------
+ *
+ *      COMPONENT "MODALBUTTONS" STYLES
+ * 
+ * ------------------------------------------
+ */
+
+const buttons = css`
+    
+  display: flex;
+  justify-content: space-between;
+  margin: 35px;
+
+`;
+
+/* ------------------------------------------
+ *
+ *                  EXPORT
+ * 
+ * ------------------------------------------
+ */
+
+export { buttons };
