+++ src/styles/components/PlacesSearch.styles.ts
@@ -1,0 +1,47 @@
+/** @jsxImportSource @emotion/react */
+
+/* ------------------------------------------
+ *
+ *                  IMPORT
+ * 
+ * ------------------------------------------
+ */
+
+import { css } from '@emotion/react';
+
+/* ------------------------------------------
+ *
+ *      COMPONENT "PLACESSEARCH" STYLES
+ * 
+ * ------------------------------------------
+ */
+
+const container = css`
+
+  display: flex;
+  padding: 5px 3px 5px 3px;
+  border: 1px solid black;
+  border-radius: 5px;
+  background: dimgrey;
+
+`;
+
+const element = css`
+
+  margin: 2px;
+
+`;
+
+/* ------------------------------------------
+ *
+ *                  EXPORT
+ * 
+ * ------------------------------------------
+ */
+
+export { 
+
+  container,
+  element
+
+};
